+++ obsplus/datasets/__init__.py
@@ -1,4 +1,4 @@
 """
 Module for loading and managing datasets in ObsPlus.
 """
-from .dataloader import DataSet
+from .dataset import DataSet
