+++ tests/test_stations/test_pd_stations.py
@@ -13,7 +13,7 @@ import pytest
 import obsplus
 from obsplus import stations_to_df
 from obsplus.constants import STATION_COLUMNS
-from obsplus.datasets.dataloader import base_path
+from obsplus.datasets.dataset import base_path
 
 KEM_PATH = base_path / "kemmerer"
 STA_COLUMNS = {"latitude", "longitude", "elevation", "start_date", "end_date"}
